feat(dsl): Enhance scope DSL and update sample app
Improves the scope<T> DSL by passing the created subcomponent instance of type T as a parameter to the whenActive lambda.
This change makes the API more intuitive and powerful, removing the boilerplate of having to re-inject the component within its own active scope. The sample application has been updated to use this new, cleaner syntax.